requirements.txt
absl_py==0.12.0
colorlog==6.4.1
fvcore==0.1.5.post20210915
numpy>=1.19.2
pandas==1.1.3
Pillow==8.4.0
spacy==3.0.7
torch==1.8.0
torchvision==0.9.0
tqdm==4.50.2
transformers==4.11.3
yacs==0.1.8
